refactorin GenderImplements

+++ src/main/java/app/disney/service/implement/GenderServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import app.disney.entitys.Gender;
+import app.disney.exceptions.NotFoundException;
 import app.disney.repository.IGenderRepository;
 import app.disney.service.IGenderService;
 
@@ -29,12 +30,15 @@ public class GenderServiceImpl implements IGenderService {
 
 	@Override
 	public Gender getGenderById(Integer id) {
-		return genderRepo.findById(id).get();
+		return genderRepo.findById(id).orElseThrow(() -> new NotFoundException("ID : " + id));
 	}
 
 	@Override
 	public void deleteGenderById(Integer id) {
-		genderRepo.deleteById(id);
+		if (genderRepo.findById(id).isPresent())
+			genderRepo.deleteById(id);
+		else
+			throw new NotFoundException("ID : " + id);
 	}
 	
 	/***BUSQUEDAS***/
